# docs/openapi/asstrack.v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "AssTrack.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/webhooks/test": {
      "post": {
        "tags": [
          "AssTrack.Api"
        ],
        "summary": "Fire a synthetic test webhook event to verify delivery configuration.",
        "operationId": "TestWebhookFire",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestWebhookRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/webhooks/subscriptions": {
      "get": {
        "tags": [
          "AssTrack.Api"
        ],
        "summary": "List event-specific webhook subscriptions.",
        "operationId": "GetWebhookSubscriptions",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "AssTrack.Api"
        ],
        "summary": "Create an event-specific webhook subscription.",
        "operationId": "CreateWebhookSubscription",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWebhookSubscriptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/webhooks/subscriptions/{id}": {
      "put": {
        "tags": [
          "AssTrack.Api"
        ],
        "summary": "Update an event-specific webhook subscription.",
        "operationId": "UpdateWebhookSubscription",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWebhookSubscriptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "AssTrack.Api"
        ],
        "summary": "Delete an event-specific webhook subscription.",
        "operationId": "DeleteWebhookSubscription",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateWebhookSubscriptionRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "isEnabled": {
            "type": "boolean"
          },
          "eventTypes": {
            "type": "string",
            "nullable": true
          },
          "targetUrl": {
            "type": "string",
            "nullable": true
          },
          "signingSecret": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TestWebhookRequest": {
        "type": "object",
        "properties": {
          "eventType": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateWebhookSubscriptionRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "isEnabled": {
            "type": "boolean"
          },
          "eventTypes": {
            "type": "string",
            "nullable": true
          },
          "targetUrl": {
            "type": "string",
            "nullable": true
          },
          "signingSecret": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "API key authentication",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
  "security": [
    { }
  ],
...
}